Video web view capability
Starting to play around with this capability. Haven't fully integrated it yet, but essentially at a much lower frame rate I can view a preview of the display sent by the server using a websocket.
Memory is getting more and more challenging though, so I may need to optimize better.