Added a page for drawing
Simple page for drawing on the LED's was added. One neat feature is when using it from a touch device, it supports tracking more than one touch event at a time. This means you can draw with two fingers if you so desire! Now granted trying to draw with your fingers on a phone screen is kind of annoying, but it definitely has a "cool" factor to it haha